// generated/src/app/api-models/hold-deny-log.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class HoldDenyLog {
  seqClaimId: number;
  lineNumber: number;
  subLineCode: string;
  startDate: Date;
  endDate: Date;
  hdIndicator: string;
  insertUser: string;
  insertDatetime: Date;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}